fix(box2d): eliminate FMA fusion so results match on amd64 and arm64
The golden bit-pattern tests were generated on arm64 and failed on the
CI amd64 runner. Root cause: the package's FMA rule only guarded
products written inline as an operand of + or -. Fusion is also legal
when an unrounded product reaches the + through a local, a struct
field, or a function parameter that is later inlined -- e.g.
MulSV's `s * v.X` flowing into mulAdd's addend, which arm64 emitted as
FMADDD and baseline amd64 did not. Disassembly showed 161 such sites.

Round products where they are formed, and additionally round the
operands the math_fma.go helpers receive so a caller's product is
rounded once the call is inlined (the smaller fix at most sites).
arm64 now emits zero FMA instructions, and the regenerated goldens
carry exactly the values amd64 was already producing.

Add TestNoFusedMultiplyAdd, which compiles the package for arm64 and
amd64/v3 and fails on any FMA instruction in the assembly. Reviewing
source for this is unreliable, so the invariant is now machine-checked
and wired into the determinism workflow ahead of the golden tests, so
a regression reports its cause rather than just "bits differ".

+++ pkg/box2d/distance.go
@@ -336,9 +336,9 @@ func solveSimplex3(s *Simplex) Vec2 {
 	// Triangle123
 	n123 := Cross(e12, e13)
 
-	d1231 := n123 * Cross(w2, w3) // upstream d123_1
-	d1232 := n123 * Cross(w3, w1) // upstream d123_2
-	d1233 := n123 * Cross(w1, w2) // upstream d123_3
+	d1231 := float64(n123 * Cross(w2, w3)) // upstream d123_1
+	d1232 := float64(n123 * Cross(w3, w1)) // upstream d123_2
+	d1233 := float64(n123 * Cross(w1, w2)) // upstream d123_3
 
 	// w1 region
 	if d122 <= 0.0 && d132 <= 0.0 {
